requirements/base.txt
Django==2.0.1
psycopg2==2.7.3.2
python-decouple==3.1
pytz
uWSGI==2.0.15
django-widget-tweaks==1.4.2 
django-wysiwyg==0.8.0 
djangorestframework==3.8.2 
djangorestframework-jwt==1.11.0 
